split progress into userdata and progress files
