+++ map.html
@@ -88,10 +88,14 @@
 	//normally, we can't call module code from non-module code.
 	//exportNamespace appends everything to window, so it can be called like normal js code.
 	import exportNamespace from './js/module-shim.mjs'
+	import * as progress from './js/progress.mjs'
+	import * as userdata from './js/userdata.mjs'
 	import * as map from './js/map.mjs';
 	import * as point from './js/map/point.mjs';
 	exportNamespace(map);
 	exportNamespace(point);
+	exportNamespace(progress);
+	exportNamespace(userdata);
 	
 	if(loadCookie("debug") == true){
 		window.debug = true;
