+++ speedrun-3.html
@@ -6,7 +6,6 @@
 	<link rel="stylesheet" type="text/css" href="guide.css"/>
 	<script src="obliviondata.js"></script>
 	<script src="base64ArrayBuffer.js"></script>
-	<script src="progress.js"></script>
 	<script src="sharing.js"></script>
 	<script src="guide.js"></script>
 	<script src="autoread.js"></script>
@@ -4741,13 +4740,17 @@ You have completed the 100% run!
 <script type="module">
 	import exportNamespace from './js/module-shim.mjs'
 	import * as common from './js/common.mjs';
+	import * as progress from './js/progress.mjs'
+	import * as userdata from './js/userdata.mjs'
 	exportNamespace(common);
-</script>
-<script>
-	if(loadCookie("debug") == true){
-		window.debug = true;
-	}
-	init();
-	makeCollapsible();
+	exportNamespace(progress);
+	exportNamespace(userdata);
+	(function(){
+		if(loadCookie("debug") == true){
+			window.debug = true;
+		}
+		init();
+		makeCollapsible();
+	}).call(window);
 </script>
 </body>
